requirements.txt
fastapi
uvicorn
requests
streamlit
pydantic
google-cloud-storage
